% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_estimates.R
\name{extract_estimates}
\alias{extract_estimates}
\title{Extract estimates from estimation object}
\usage{
extract_estimates(o, digits = 3, coef.extractor = coef,
  vcov.extractor = vcov)
}
\description{
Extract estimates from estimation object
}

